+++ test/gateway/eip20_cogateway/progress_redeem_with_proof.js
@@ -509,68 +509,6 @@ contract('EIP20CoGateway.progressRedeemWithProof() ', (accounts) => {
     );
   });
 
-  it('should return penalty to redeemer when the message status in source is '
-    + 'declared revocation and in the target is progressed', async () => {
-    facilitator = accounts[8];
-    redeemer = redeemParams.redeemer;
-
-    await web3.eth.sendTransaction(
-      {
-        to: eip20CoGateway.address,
-        from: facilitator,
-        value: penaltyAmount,
-      },
-    );
-
-    await eip20CoGateway.setOutboxStatus(
-      redeemParams.messageHash,
-      MessageStatusEnum.DeclaredRevocation,
-    );
-
-    const initialFacilitatorEthBalance = await Utils.getBalance(facilitator);
-    const initialRedeemerEthBalance = await Utils.getBalance(redeemer);
-    const initialCoGatewayEthBalance = await Utils.getBalance(eip20CoGateway.address);
-
-    await setStorageRoot();
-
-    const tx = await eip20CoGateway.progressRedeemWithProof(
-      redeemParams.messageHash,
-      proofData.gateway.progress_unstake.proof_data.storageProof[0].serializedProof,
-      new BN(proofData.gateway.progress_unstake.proof_data.block_number, 16),
-      MessageStatusEnum.Progressed,
-      { from: facilitator },
-    );
-
-    const finalFacilitatorEthBalance = await Utils.getBalance(facilitator);
-    const finalRedeemerEthBalance = await Utils.getBalance(redeemer);
-    const finalCoGatewayEthBalance = await Utils.getBalance(eip20CoGateway.address);
-
-    const expectedFinalFacilitatorETHBalance = initialFacilitatorEthBalance
-      .add(bountyAmount)
-      .subn(tx.receipt.gasUsed);
-
-    const expectedFinalRedeemerETHBalance = initialRedeemerEthBalance
-      .add(penaltyAmount);
-
-    assert.strictEqual(
-      finalFacilitatorEthBalance.eq(expectedFinalFacilitatorETHBalance),
-      true,
-      `Facilitator's base token balance ${finalFacilitatorEthBalance.toString(10)} should be equal to ${expectedFinalFacilitatorETHBalance.toString(10)}`,
-    );
-
-    assert.strictEqual(
-      finalRedeemerEthBalance.eq(expectedFinalRedeemerETHBalance),
-      true,
-      `Redeemer's base token balance ${finalRedeemerEthBalance.toString(10)} should be equal to ${expectedFinalRedeemerETHBalance.toString(10)}`,
-    );
-
-    assert.strictEqual(
-      finalCoGatewayEthBalance.eq(initialCoGatewayEthBalance.sub(bountyAmount).sub(penaltyAmount)),
-      true,
-      `CoGateway's base token balance ${finalCoGatewayEthBalance.toString(10)} should be equal to ${initialCoGatewayEthBalance.sub(bountyAmount).sub(penaltyAmount)}.`,
-    );
-  });
-
   it('should decrease token supply for utility token', async () => {
     const initialTotalSupply = await utilityToken.totalSupply.call();
     const initialCoGatewayBalance = await utilityToken.balanceOf(eip20CoGateway.address);